milieu rapport

+++ code/ville.h
@@ -3,13 +3,13 @@
 
 typedef struct Ville_t Ville;
 
-Ville *creer_ville(char *nom, int x, int y);
+Ville *creer_ville(char *nom, float x, float y);
 
 void detruit_ville(Ville *ville);
 
-int get_x_ville(Ville *ville);
+float get_x_ville(Ville *ville);
 
-int get_y_ville(Ville *ville);
+float get_y_ville(Ville *ville);
 
 char *get_nom_ville(Ville *ville);
 
@@ -17,7 +17,7 @@ char *get_specialite_ville(Ville *ville);
 
 void set_specialite_ville(Ville *ville, char *specialite);
 
-double distance_entre_2_villes(Ville *ville1, Ville *ville2);
+float distance_entre_2_villes(Ville *ville1, Ville *ville2);
 
 int size_ville(void);
